Updated components Updated CreateCityForm.vue, CreateGuestForm.vue, GuestList.vue, cities.json, cityStore.js, guestStore.js, SideBarLayout.vue, CityPage.vue. Added SideBarList.vue
