+++ src/stores/countryStore.js
@@ -9,17 +9,11 @@ export const useCountryStore = defineStore('CountryStore', {
     actions: {
         async loadAllCountries() {
             this.countries = (await import("../data/countries.json")).default;
-        },
-        storeCountry(country) {
-            this.countries.push({
-                id: this.countries.length + 1,
-                ...country
-            });
         }
     },
     getters: {
         getAllCountries(state) {
             return  [...state.countries];
-        },
+        }
     }
 })
